Fix revision

+++ server/controllers/dataset-controller.js
@@ -1,4 +1,4 @@
-const { fetchDataFromDatabase } = require("../nlp");
+const { fetchDataFromDatabase, manager } = require("../nlp");
 const db = require("../models");
 
 exports.getAll = async (req, res) => {
@@ -113,7 +113,22 @@ exports.delete = async (req, res) => {
     try {
         const { id } = req.params;
 
-        const dataset = await db.Dataset.destroy({
+        const dataset = await db.Dataset.findOne({
+            where: {
+                id
+            }
+        });
+
+        if (!dataset) {
+            return res.status(404).send({
+                success: false,
+                message: "Dataset not found",
+            });
+        }
+
+        manager.removeDocument('id', dataset.utterance, dataset.intent);
+
+        await db.Dataset.destroy({
             where: {
                 id
             }
@@ -164,7 +179,17 @@ exports.deleteIntent = async (req, res) => {
     try {
         const { intent } = req.body;
 
-        const dataset = await db.Dataset.destroy({
+        const dataset = await db.Dataset.findAll({
+            where: {
+                intent
+            }
+        });
+
+        for (const item of dataset) {
+            manager.removeDocument('id', item.utterance, intent);
+        }
+
+        await db.Dataset.destroy({
             where: {
                 intent
             }
